Bug Fix in Following and Position Output When Stopped
Fixed a bug where an agent would follow an agent every time step when a
follow decision was made. This was similar to the initiation bug in the
previous commit.
Fixed a bug in the positional output where there would be oscillating in
the agent's location after they reached their destination.